get common edge
