+++ rust/kcl-lib/src/std/args.rs
@@ -144,7 +144,7 @@ impl Args {
             KclError::Type(KclErrorDetails {
                 source_ranges: vec![self.source_range],
                 message: format!(
-                    "The optional arg {label} was given, but it was the wrong type. It should be type {} but it was {}",
+                    "The arg {label} was given, but it was the wrong type. It should be type {} but it was {}",
                     type_name::<T>(),
                     arg.value.human_friendly_type(),
                 ),
@@ -969,6 +969,22 @@ impl<'a> FromKclValue<'a> for TagNode {
 impl<'a> FromKclValue<'a> for TagIdentifier {
     fn from_kcl_val(arg: &'a KclValue) -> Option<Self> {
         arg.get_tag_identifier().ok()
+    }
+}
+
+impl<'a> FromKclValue<'a> for Vec<TagIdentifier> {
+    fn from_kcl_val(arg: &'a KclValue) -> Option<Self> {
+        match arg {
+            KclValue::HomArray { value, .. } => {
+                let tags = value.iter().map(|v| v.get_tag_identifier().unwrap()).collect();
+                Some(tags)
+            }
+            KclValue::MixedArray { value, .. } => {
+                let tags = value.iter().map(|v| v.get_tag_identifier().unwrap()).collect();
+                Some(tags)
+            }
+            _ => None,
+        }
     }
 }
