+++ rust/kcl-lib/src/std/fillet.rs
@@ -3,14 +3,10 @@
 use anyhow::Result;
 use indexmap::IndexMap;
 use kcl_derive_docs::stdlib;
-use kcmc::{
-    each_cmd as mcmd, length_unit::LengthUnit, ok_response::OkModelingCmdResponse, shared::CutType,
-    websocket::OkWebSocketResponseData, ModelingCmd,
-};
+use kcmc::{each_cmd as mcmd, length_unit::LengthUnit, shared::CutType, ModelingCmd};
 use kittycad_modeling_cmds as kcmc;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use uuid::Uuid;
 
 use crate::{
     errors::{KclError, KclErrorDetails},
@@ -197,250 +193,6 @@ async fn inner_fillet(
     }
 
     Ok(solid)
-}
-
-/// Get the opposite edge to the edge given.
-pub async fn get_opposite_edge(exec_state: &mut ExecState, args: Args) -> Result<KclValue, KclError> {
-    let tag: TagIdentifier = args.get_data()?;
-
-    let edge = inner_get_opposite_edge(tag, exec_state, args.clone()).await?;
-    Ok(KclValue::Uuid {
-        value: edge,
-        meta: vec![args.source_range.into()],
-    })
-}
-
-/// Get the opposite edge to the edge given.
-///
-/// ```no_run
-/// exampleSketch = startSketchOn('XZ')
-///   |> startProfileAt([0, 0], %)
-///   |> line(end = [10, 0])
-///   |> angledLine({
-///     angle = 60,
-///     length = 10,
-///   }, %)
-///   |> angledLine({
-///     angle = 120,
-///     length = 10,
-///   }, %)
-///   |> line(end = [-10, 0])
-///   |> angledLine({
-///     angle = 240,
-///     length = 10,
-///   }, %, $referenceEdge)
-///   |> close()
-///
-/// example = extrude(exampleSketch, length = 5)
-///   |> fillet(
-///     radius = 3,
-///     tags = [getOppositeEdge(referenceEdge)],
-///   )
-/// ```
-#[stdlib {
-    name = "getOppositeEdge",
-}]
-async fn inner_get_opposite_edge(tag: TagIdentifier, exec_state: &mut ExecState, args: Args) -> Result<Uuid, KclError> {
-    if args.ctx.no_engine_commands().await {
-        return Ok(exec_state.next_uuid());
-    }
-    let face_id = args.get_adjacent_face_to_tag(exec_state, &tag, false).await?;
-
-    let id = exec_state.next_uuid();
-    let tagged_path = args.get_tag_engine_info(exec_state, &tag)?;
-
-    let resp = args
-        .send_modeling_cmd(
-            id,
-            ModelingCmd::from(mcmd::Solid3dGetOppositeEdge {
-                edge_id: tagged_path.id,
-                object_id: tagged_path.sketch,
-                face_id,
-            }),
-        )
-        .await?;
-    let OkWebSocketResponseData::Modeling {
-        modeling_response: OkModelingCmdResponse::Solid3dGetOppositeEdge(opposite_edge),
-    } = &resp
-    else {
-        return Err(KclError::Engine(KclErrorDetails {
-            message: format!("mcmd::Solid3dGetOppositeEdge response was not as expected: {:?}", resp),
-            source_ranges: vec![args.source_range],
-        }));
-    };
-
-    Ok(opposite_edge.edge)
-}
-
-/// Get the next adjacent edge to the edge given.
-pub async fn get_next_adjacent_edge(exec_state: &mut ExecState, args: Args) -> Result<KclValue, KclError> {
-    let tag: TagIdentifier = args.get_data()?;
-
-    let edge = inner_get_next_adjacent_edge(tag, exec_state, args.clone()).await?;
-    Ok(KclValue::Uuid {
-        value: edge,
-        meta: vec![args.source_range.into()],
-    })
-}
-
-/// Get the next adjacent edge to the edge given.
-///
-/// ```no_run
-/// exampleSketch = startSketchOn('XZ')
-///   |> startProfileAt([0, 0], %)
-///   |> line(end = [10, 0])
-///   |> angledLine({
-///     angle = 60,
-///     length = 10,
-///   }, %)
-///   |> angledLine({
-///     angle = 120,
-///     length = 10,
-///   }, %)
-///   |> line(end = [-10, 0])
-///   |> angledLine({
-///     angle = 240,
-///     length = 10,
-///   }, %, $referenceEdge)
-///   |> close()
-///
-/// example = extrude(exampleSketch, length = 5)
-///   |> fillet(
-///     radius = 3,
-///     tags = [getNextAdjacentEdge(referenceEdge)],
-///   )
-/// ```
-#[stdlib {
-    name = "getNextAdjacentEdge",
-}]
-async fn inner_get_next_adjacent_edge(
-    tag: TagIdentifier,
-    exec_state: &mut ExecState,
-    args: Args,
-) -> Result<Uuid, KclError> {
-    if args.ctx.no_engine_commands().await {
-        return Ok(exec_state.next_uuid());
-    }
-    let face_id = args.get_adjacent_face_to_tag(exec_state, &tag, false).await?;
-
-    let id = exec_state.next_uuid();
-    let tagged_path = args.get_tag_engine_info(exec_state, &tag)?;
-
-    let resp = args
-        .send_modeling_cmd(
-            id,
-            ModelingCmd::from(mcmd::Solid3dGetNextAdjacentEdge {
-                edge_id: tagged_path.id,
-                object_id: tagged_path.sketch,
-                face_id,
-            }),
-        )
-        .await?;
-
-    let OkWebSocketResponseData::Modeling {
-        modeling_response: OkModelingCmdResponse::Solid3dGetNextAdjacentEdge(adjacent_edge),
-    } = &resp
-    else {
-        return Err(KclError::Engine(KclErrorDetails {
-            message: format!(
-                "mcmd::Solid3dGetNextAdjacentEdge response was not as expected: {:?}",
-                resp
-            ),
-            source_ranges: vec![args.source_range],
-        }));
-    };
-
-    adjacent_edge.edge.ok_or_else(|| {
-        KclError::Type(KclErrorDetails {
-            message: format!("No edge found next adjacent to tag: `{}`", tag.value),
-            source_ranges: vec![args.source_range],
-        })
-    })
-}
-
-/// Get the previous adjacent edge to the edge given.
-pub async fn get_previous_adjacent_edge(exec_state: &mut ExecState, args: Args) -> Result<KclValue, KclError> {
-    let tag: TagIdentifier = args.get_data()?;
-
-    let edge = inner_get_previous_adjacent_edge(tag, exec_state, args.clone()).await?;
-    Ok(KclValue::Uuid {
-        value: edge,
-        meta: vec![args.source_range.into()],
-    })
-}
-
-/// Get the previous adjacent edge to the edge given.
-///
-/// ```no_run
-/// exampleSketch = startSketchOn('XZ')
-///   |> startProfileAt([0, 0], %)
-///   |> line(end = [10, 0])
-///   |> angledLine({
-///     angle = 60,
-///     length = 10,
-///   }, %)
-///   |> angledLine({
-///     angle = 120,
-///     length = 10,
-///   }, %)
-///   |> line(end = [-10, 0])
-///   |> angledLine({
-///     angle = 240,
-///     length = 10,
-///   }, %, $referenceEdge)
-///   |> close()
-///
-/// example = extrude(exampleSketch, length = 5)
-///   |> fillet(
-///     radius = 3,
-///     tags = [getPreviousAdjacentEdge(referenceEdge)],
-///   )
-/// ```
-#[stdlib {
-    name = "getPreviousAdjacentEdge",
-}]
-async fn inner_get_previous_adjacent_edge(
-    tag: TagIdentifier,
-    exec_state: &mut ExecState,
-    args: Args,
-) -> Result<Uuid, KclError> {
-    if args.ctx.no_engine_commands().await {
-        return Ok(exec_state.next_uuid());
-    }
-    let face_id = args.get_adjacent_face_to_tag(exec_state, &tag, false).await?;
-
-    let id = exec_state.next_uuid();
-    let tagged_path = args.get_tag_engine_info(exec_state, &tag)?;
-
-    let resp = args
-        .send_modeling_cmd(
-            id,
-            ModelingCmd::from(mcmd::Solid3dGetPrevAdjacentEdge {
-                edge_id: tagged_path.id,
-                object_id: tagged_path.sketch,
-                face_id,
-            }),
-        )
-        .await?;
-    let OkWebSocketResponseData::Modeling {
-        modeling_response: OkModelingCmdResponse::Solid3dGetPrevAdjacentEdge(adjacent_edge),
-    } = &resp
-    else {
-        return Err(KclError::Engine(KclErrorDetails {
-            message: format!(
-                "mcmd::Solid3dGetPrevAdjacentEdge response was not as expected: {:?}",
-                resp
-            ),
-            source_ranges: vec![args.source_range],
-        }));
-    };
-
-    adjacent_edge.edge.ok_or_else(|| {
-        KclError::Type(KclErrorDetails {
-            message: format!("No edge found previous adjacent to tag: `{}`", tag.value),
-            source_ranges: vec![args.source_range],
-        })
-    })
 }
 
 pub(crate) fn default_tolerance(units: &UnitLength) -> f64 {
