+++ rust/kcl-lib/src/std/mod.rs
@@ -7,6 +7,7 @@ pub mod assert;
 pub mod axis_or_reference;
 pub mod chamfer;
 pub mod convert;
+pub mod edge;
 pub mod extrude;
 pub mod fillet;
 pub mod helix;
@@ -107,9 +108,10 @@ lazy_static! {
         Box::new(crate::std::array::Pop),
         Box::new(crate::std::chamfer::Chamfer),
         Box::new(crate::std::fillet::Fillet),
-        Box::new(crate::std::fillet::GetOppositeEdge),
-        Box::new(crate::std::fillet::GetNextAdjacentEdge),
-        Box::new(crate::std::fillet::GetPreviousAdjacentEdge),
+        Box::new(crate::std::edge::GetOppositeEdge),
+        Box::new(crate::std::edge::GetNextAdjacentEdge),
+        Box::new(crate::std::edge::GetPreviousAdjacentEdge),
+        Box::new(crate::std::edge::GetCommonEdge),
         Box::new(crate::std::helix::Helix),
         Box::new(crate::std::helix::HelixRevolutions),
         Box::new(crate::std::shell::Shell),
